refactor(db): file-organization sweep over packages/db
The three shapes its services declared beside the code that reads them — the blob-listing options, the message-creation rejection and the permissions overload set — are model files now, and the listing options are exported so the server's profile-image sweep can take the same shape instead of restating it. The two Azure error predicates every consumer of which sits under table/ move there, per the sole-consumer rule.
