+++ packages/db/src/models/message/moderation/MessageCreationRejection.ts
@@ -1,0 +1,8 @@
+import type { MessageCreationRejectionType, RoomFilterInMessage } from "@esposter/db-schema";
+
+export type MessageCreationRejection =
+  | {
+      filter: Pick<RoomFilterInMessage, "action" | "timeoutDurationMs">;
+      type: MessageCreationRejectionType.WordFilter;
+    }
+  | { type: Exclude<MessageCreationRejectionType, MessageCreationRejectionType.WordFilter> };
