+++ packages/db/src/services/azure/table/checkIsNotFound.ts
