+++ packages/db/src/services/azure/table/getTableClient.ts
@@ -1,6 +1,6 @@
 import type { AzureTable, AzureTableEntityMap, CustomTableClient } from "@esposter/db-schema";
 
-import { checkIsConflict } from "#src/services/azure/checkIsConflict";
+import { checkIsConflict } from "#src/services/azure/table/checkIsConflict";
 import { createProvisionedClientCache } from "#src/services/azure/createProvisionedClientCache";
 import { TableClient } from "@azure/data-tables";
 import { getResultAsync, noop } from "@esposter/shared";
